+++ inst/srcjs/inject.js
@@ -1,0 +1,1 @@
+Shiny.addCustomMessageHandler('shinyjs-inject', function(content) {$("head").append(content);});